+++ src/test/resources/edgeCases/createModel.php
@@ -5,5 +5,6 @@ class User extends \Illuminate\Database\Eloquent\Model {
 }
 
 User::create([
+    'email' => 'email@email.com',
     '<caret>' => 1,
 ]);