+++ src/main/kotlin/dev/ekvedaras/laravelquery/reference/ColumnReferenceProvider.kt
@@ -13,11 +13,12 @@ import dev.ekvedaras.laravelquery.models.DbReferenceExpression
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.canHaveColumnsInArrayValues
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isBuilderMethodForColumns
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isColumnIn
-import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsidePhpArrayOrValue
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInsideRegularFunction
 import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.isInteresting
+import dev.ekvedaras.laravelquery.utils.LaravelUtils.Companion.shouldCompleteOnlyColumns
 import dev.ekvedaras.laravelquery.utils.MethodUtils
 import dev.ekvedaras.laravelquery.utils.PsiUtils.Companion.containsVariable
+import dev.ekvedaras.laravelquery.utils.PsiUtils.Companion.isArrayValue
 
 class ColumnReferenceProvider : PsiReferenceProvider() {
     companion object {
@@ -41,23 +42,27 @@ class ColumnReferenceProvider : PsiReferenceProvider() {
         val target = DbReferenceExpression(element, DbReferenceExpression.Companion.Type.Column)
         var references = arrayOf<PsiReference>()
 
-        target.schema.parallelStream().forEach { references += SchemaPsiReference(target, it) }
-        target.table.parallelStream().forEach {
-            references += TableOrViewPsiReference(target, it)
+        if (!method.shouldCompleteOnlyColumns()) {
+            target.schema.parallelStream().forEach { references += SchemaPsiReference(target, it) }
 
-            val alias = target.aliases[it.name]
-            if (alias != null) {
-                references += TableAliasPsiReference(
-                    element,
-                    if (target.ranges.size >= 2 && target.schema.isNotEmpty()) {
-                        target.ranges[1]
-                    } else {
-                        target.ranges.first()
-                    },
-                    alias.second
-                )
+            target.table.parallelStream().forEach {
+                references += TableOrViewPsiReference(target, it)
+
+                val alias = target.aliases[it.name]
+                if (alias != null) {
+                    references += TableAliasPsiReference(
+                        element,
+                        if (target.ranges.size >= 2 && target.schema.isNotEmpty()) {
+                            target.ranges[1]
+                        } else {
+                            target.ranges.first()
+                        },
+                        alias.second
+                    )
+                }
             }
         }
+
         target.column.parallelStream()
             .filter {
                 target.tablesAndAliases.isEmpty() ||
@@ -80,7 +85,7 @@ class ColumnReferenceProvider : PsiReferenceProvider() {
             !method.isBuilderMethodForColumns() ||
             !element.isColumnIn(method, allowArray) ||
             element.isInsideRegularFunction() ||
-            (element.isInsidePhpArrayOrValue() && !method.canHaveColumnsInArrayValues()) ||
+            ((element.parent?.isArrayValue() ?: false) && !method.canHaveColumnsInArrayValues()) ||
             !method.isInteresting(project)
     }
 }