+++ src/test/resources/edgeCases/createModelKeyAndValueAsTableName.php
@@ -1,0 +1,9 @@
+<?php
+
+class User extends \Illuminate\Database\Eloquent\Model {
+
+}
+
+User::create([
+    'users' => 'users',
+]);