+++ src/main/kotlin/dev/ekvedaras/laravelquery/utils/PsiUtils.kt
@@ -14,7 +14,7 @@ import org.jetbrains.annotations.NotNull
 @Suppress("MagicNumber")
 object ElementTypes {
     val PhpArray = listOf(1386, 1889)
-    val ArrayValues = listOf(805, 816, 1405)
+    const val ArrayValues = "Array value"
 }
 
 class PsiUtils private constructor() {
@@ -23,7 +23,7 @@ class PsiUtils private constructor() {
         fun CompletionParameters.containsVariable(): Boolean = this.position.containsVariable()
         fun String.containsAlias(): Boolean = this.contains(" as ")
         fun PsiElement.isPhpArray(): Boolean = ElementTypes.PhpArray.contains(this.typeAsInt())
-        fun PsiElement.isArrayValue(): Boolean = ElementTypes.ArrayValues.contains(this.typeAsInt())
+        fun PsiElement.isArrayValue(): Boolean = this.elementType.toString() === ElementTypes.ArrayValues
         fun String.unquoteAndCleanup() = this.replace("IntellijIdeaRulezzz", "").trim('\'', '"').trim()
         fun Variable.references(): @NotNull Query<PsiReference> =
             ReferencesSearch.search(this.originalElement)