+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 - Do not inspect for tables in `create`, `update`, `fill`, etc. methods (#38)
 - Do not complete columns in `create`, `update`, `fill`, etc. method array values. Only in keys (#38)
+- Do not look for column and table references in `create`, `update`, `fill`, etc. method array values. Only in keys (#38)
 
 ## [2.0.0]
 ### Added