+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # Laravel Query Changelog
 
 ## [Unreleased]
+
+## [2.0.1]
 ### Fixed
 - Do not inspect for tables in `create`, `update`, `fill`, etc. methods (#38)
 - Do not complete columns in `create`, `update`, `fill`, etc. method array values. Only in keys (#38)